feat(Subtick): allow modules to have a blit and a logic loop
By having a single loop we force the user interaction to run as slow as
the blitting. This causes the game to feel a bit weird as every command
takes a short but noticeable time to run.

Now we are allowing modules to split their loop into the update_all main
loop and the subtick that runs on every frame.