refactor: clean up main.py
With this change, I don't yet understand why, some sound issues seem to
be  fixed.

Previous to this the new main menu music was having some noisy sounds in
high notes and the steps and kiss sounds wore not working.

I won't consider this a fix for now as it may be that those were issues
with my environment.